[FAILS] Add a test to show issue with subdomains as additional permissions
According to the documentation, `strictOrigins` is only concerned with
widening of permissions, not additional permissions which are narrower
than a pattern already in the manifest origins.

However, if `https://subdomain.hassubdomains.com/*` is granted by the
user as an optional permission on top of `https://*.hassubdomains.com/*`,
this does not change the effective set of permissions.

So to be consistent with the documentation, when `strictOrigins` is
`true`, the additional subdomain pattern should not be included in the
return value of `extractAdditionalPermissions()`.

Adding a test based on this expectation causes a failure, showing the
current problem.  So either the code should be changed, or the
documentation for `strictOrigins` expanded to clarify that it also
catches narrower permissions as well as wider ones.